+++ src/main/java/techcourse/myblog/domain/dto/LoginRequestDto.java
@@ -1,4 +1,4 @@
-package techcourse.myblog.domain;
+package techcourse.myblog.domain.dto;
 
 public class LoginRequestDto {
