+++ src/main/java/techcourse/myblog/web/UserController.java
@@ -8,6 +8,9 @@ import org.springframework.ui.Model;
 import org.springframework.validation.BindingResult;
 import org.springframework.web.bind.annotation.*;
 import techcourse.myblog.domain.*;
+import techcourse.myblog.domain.dto.LoginRequestDto;
+import techcourse.myblog.domain.dto.MyPageRequestDto;
+import techcourse.myblog.domain.dto.UserRequestDto;
 
 import javax.servlet.http.HttpSession;
 import javax.transaction.Transactional;
